pyproject.toml
[project]
name = "anint"
version = "0.0.1"
dependencies = [
    "pyyaml",
]
authors = [
    { name = "Ankhbold Batbaatar", email = "ankhasgreens@gmail.com" },
]
maintainers = [
    { name = "Ankhbold Batbaatar", email = "ankhasgreens@gmail.com" },
]
description = "Ankha's Internationalization and Localization for Python."
readme = "README.md"
requires-python = ">=3.11"
license = { text = "MIT" }
keywords = [ "Ankha's Greens", "Anint", "I18n", "Internationalization", "Localization", "Translation" ]
classifiers = [
    "License :: OSI Approved :: MIT License",
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Topic :: Software Development :: Build Tools",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Programming Language :: Python :: 3.13",
    "Operating System :: OS Independent",
    "Framework :: tox",
    "Typing :: Typed",
]

[project.urls]
Homepage = ""
Issues = ""

[build-system]
requires = ["setuptools>=61.0"]
build-backend = "setuptools.build_meta"

[tool.setuptools.package-data]
anint = [ "py.typed" ]

[tool.pytest.ini_options]
minversion = "6.0"
testpaths = [ "tests" ]

[tool.mypy]
mypy_path = "src"
check_untyped_defs = true
